- renamed component - added mobile responsive css
