+++ src/components/flexWrapper/flexWrapper.scss
@@ -6,6 +6,10 @@
 
 .row {
   flex-direction: row;
+
+  @include for-phone-only {
+    flex-direction: column;
+  }
 }
 
 .column {
