+++ src/components/index.ts
@@ -45,4 +45,4 @@ export * from "./iconButton";
 export * from "./EnumValues";
 export * from "./g2bDatePicker";
 export * from "./searchBarSingleDropdown";
-export * from "./flex";
+export * from "./flexWrapper";
