+++ src/stories/utils.tsx
@@ -35,7 +35,7 @@ export enum CategoryName {
   Buttons = "Buttons",
   Cards = "Cards",
   Dropdown = "Dropdown",
-  Flex = "Flex",
+  FlexWrapper = "FlexWrapper",
   Icons = "Icons",
   TextFields = "Text-Fields",
   Links = "Links",
